Install Node dependencies for bip39
This commit installs packages that bip39 depends on that is not
available in React Native. It then uses babel-plugin-rewrite-require
to rewrite the requires for these dependencies to use the React
Native versions instead. And last it exposes some global variables
such as Buffer and process using globals.js.

This solution is based on the idea of ReactNativify:
https://github.com/philikon/ReactNativify

But has been made much simpler thanks to this thread:
https://github.com/philikon/ReactNativify/issues/4