Show the Launch Screen until React has finished loading
The Launch Screen is only shown until the app has loaded, not until
React has loaded. This results in a white flash between the Launch
Screen and the initial view of the app. This commit fixes that by
showing the Launch Screen until React also has finished loading.

The solution was found here:
https://github.com/facebook/react-native/issues/1402#issuecomment-130236507