Make the Splash Screen look like the Launch Screen
The Splash Screen also shows a loading animation.

The reason for the Splash Screen is to show during the app
is initializing.

Note:
The logo wasn't displaying at exactly the same position as in the
Launch Screen so I had to use marginLeft to adjust it. Might be
worth investigating and fixing it properly.