Rename utxos/init action to utxos/update
This commit renames the init action to update. This is because the
init action will be used every time the utxo set needs to be updated.
This is of course not optimal for performance but it reduces the scope
and it decreases the risk for errors (e.g. the utxo set being out of
sync).