Use FlatList instead of react-native-swiper
react-native-swiper is no longer maintained and contains a lot of bugs.
FlatList is a standard React Native component that seems to be working
exactly the same for this use-case but without the extra dependency.

At the same time, this commit also makes sure to pause the camera
preview when it's not active or being swiped to.