Force app to use built-in keyboard when entering recovery phrase
Force the app to use the built-in keyboard instead of any 3rd-party
ones when entering the recovery phrase. The reason is to prevent
3rd-party keyboards from leaking the recovery phrase.

This is done by setting `secureTextEntry` to `true` on the text
inputs. This also causes the entered text to be obscured, which
is a good side-effect (although slightly less user-friendly).