Add an illustration and text when wallet is empty
This view is only shown when the wallet doesn't contain any
transactions.

The illustration was purchased from kit8 and then slightly
modified. It's only a simple license so it will need to be
extended or replaced before the app is released. A set of
new and unique illustrations is planned for the app.