Select UTXOs and create a transaction
Uses bitcoinjs to create the transaction and coinselect to select
the UTXOs to be used as inputs.

TODO: Derive the amount and fee from the created transaction to make
sure the UI reflects the built transaction. This is to prevent
displaying an incorrect value in case there are any bugs in the coin
selection algorithm or transaction builder.