Round fee rate to nearest satoshi
Previously this was done just before creating the transaction
because the coin selection algorithm that is being used doesn't
support decimal fee rates. But now this is put in a central place
so that all estimated fees will be the same - both when changing
fee settings and when creating a new transaction.