Add React Native Navigation
This commit adds the React Native Navigation library to the app.
Currently, React Navigation is being used but the plan is to
eventually migrate to React Native Navigation instead.

The reason why this is added now is because it's at the moment
the only way to add Peek and Pop to a React Native app. It will
be used to show transaction details.

This commit introduces a visual bug that causes a white flash
between the LaunchScreen and the SplashScreen. There's a PR for
it in wix/react-native-navigation that will hopefully be merged
soon.