Revert "Add React Native Navigation"
This reverts commit a5b4b0842ea05bad2223f2739e9e052e62f48436.

Decided not to go with Peek and Pop (3D Touch) and React Native
Navigation:

Unfortunately it's very glitchy when scrolling or swiping at the
same time as peeking/previewing. Disabling scroll doesn't help
and even makes it worse at some times.

It also requires the use of React Native Navigation which comes
with its own set of challenges - e.g. showing the LaunchScreen
until React has finished loading, or having a custom Back icon
without disabling back gestures.

This can be revisited in the future and the work has been saved
in the branch `react-native-navigation-peek-and-pop`.