Format fiat currencies using their symbol when possible
The actual currency amount is formatted according to the user's
region locale while the local currency symbol is used for some
currencies, such as USD, EUR, SEK, and more. When there is no
symbol, the currency code will be used instead.

The reason for why Intl.NumberFormat isn't used to format the
amount with the currency symbol is because it requires the local
locale in order for the symbol to be correct. Without the Swedish
locale, SEK would be displayed as SKr instead of kr.