Use another cropping component to crop profile pictures
The built in cropping component by Apple is very buggy. E.g. on
iPhone X it is missing an overlay at the top. And it doesn't
support circular overlays. But worst of all, it returns an
image that is incorrectly cropped (an offset of about 20px) and
the dimensions is sometimes off by a pixel. It also seems to have
problems on iPad.

<https://stackoverflow.com/questions/48785503/uiimagepickercontroller-cropping-image-rect-is-not-correct>

The solution was to use another cropping component that's included
in another image picker. It is now a bit slower as it needs to load
the other cropping screen, but it works much better.