Store pine credentials in state to improve performance
Loading the mnemonic from the key chain and then deriving the
child key pair used to authenticate pine requests is very slow
and freezes the whole app and its UI. This causes a very poor
user experience. Other pine actions are also very slow. This
commit addresses that by storing the pine credentials in the
state. It's less secure but it's important to note that these
credentials _cannot_ be used to sign bitcoin transactions.