Load users in parallel to prevent unresponsive API calls from stacking up
The request timeout is currently set to 5 seconds, and 5 seconds for
each unresponsive call could quickly add up to more or less preventing
the user from syncing at all, i.e. DOSing the app. Making the requests
in parallel instead of serial prevents this from happening as the total
time will not be much more than 5 seconds even if there are multiple
unresponsive API calls. In the future it would be nice to inform the
user about unresponsive servers so that they can be ignored or its
contacts deleted (if persistent).