Reserve UTXOs that's been used in a payment
Note: There is a problem with this. E.g. if the user only has one
UTXO and uses it in a transaction, the user won't be able to send
any more transactions until the recipient has broadcasted it to the
bitcoin network and there are new UTXOs in form of change.

A solution would be to broadcast the transaction instead of sending
it in the message, but that would be risky since the recipient can't
verify it before it's broadcasted. Another solution would be to have
the options to withdraw a payment in favour of another.