Add some bottom padding to the beta notice label
Otherwise there will be no padding between the label and the keyboard.
However, using the Swiftkey keyboard there will be too much padding.
This could probably be fixed by using a flex layout instead of absolute
positioning it, but since it's just a temporary label it might not be
worth spending time on that now. But it's worth noting for the future.