Return sync promise if app already is syncing
This fixes a bug where the app didn't show a contact request
even though it was opened by a notification. The reason was
that the app was already syncing because of the app getting
active and then when the app was going to be synced by the
notification, an empty promise resolved early before the
sync was actually complete.

This commit fixes that by storing the sync promise and
returning that if the sync action is called while a sync
already is in progress.