Sync profiles when app becomes active
Previously, contacts's profiles were only synced at app start,
but since the app might not restart for a long time and just
be in the background, profiles will now also be updated every
time the app becomes active. However, profiles will maximum
be updated once every minute.

This change required the Conversation Screen to take a contact
ID instead of the actual contact object so that it can be updated
when the state changes.