Save intermediate addresses if address index is higher than current index
This fixes the following bug:

Cause: Address was not in already generated addresses but still below the
current index. This caused it to be missed because it was not in the
existing addresses and not in the new.

Fix: When receiving new payments and scanning the addresses, don't only
save the new address but also all other generated addresses so the sequence
is not broken.

Bonus: This is also an optimization as unused addresses will be stored and
used later without the need to be regenerated.