Upgrade react-native-camera to the latest 1.x
This fixes a couple of cases where the app crashes when scanning
invalid QR codes. There is still one case left that it doesn't
fix so it will have to be fixed in another way.

Upgrading to 2.x is not desirable because it is using Firebase/Google
to do the QR code scanning.