Always show amounts from bitcoin URIs in default BTC unit
The feature that remembers the denomination choice introduced a
bug that caused amounts scanned from a QR code to be converted to
fiat if that was the last used denomination. To prevent users from
sending incorrect amounts, the initial currency/unit will now be
BTC/default unit when scanning QR codes with btc values.