Make it possible to "cancel" a payment
Outgoing transactions that has not yet been broadcasted by its
recipients will lock up change which will only be unlocked once
the transaction has been broadcasted (or once the reservations
have expired after 2 days). To prevent being unable to send a
payment due to too much being locked up as change, add the option
to cancel sent payments that has not yet been broadcasted.

This does not guarantee that the payment will not be completed
as the recipient might still have received the transaction
without just broadcasting it (e.g. with a modified client).
To invalidate the transaction, the user should make a new
payment to themselves (or someones else) with the same
amount to be sure that one of the same inputs will be used.

This could be a feature in the future where the app asks
the user if they want to invalidate the transaction for
a fee (transaction fee).