Normalize BTC amounts when doing floating-point arithmetic
JavaScript sometimes lose precision when performing floating-point arithmetic.
The precision loss is insignificant to BTC amounts which has a precision of 8
decimals, but it is a problem when comparing numbers, e.g. 0.7 + 0.1 < 0.8 => true.
It could also become a rounding error when summing up large amounts of values.

To prevent this, always normalize BTC amounts to have a precision of 8 decimals
after a new BTC amount has been produced by floating-point arithmetic.
If the amount has more than 8 decimals, the 8th decimal will be rounded. Since
the precision loss is so low, any errors will be rounded to the desired number.