Prevent recovering recovery key with no account
Previously the app would just ask the user to pick a username if
the recovery key didn't have any existing account attached to it.
That could be confusing if the user entered the incorrect recovery
key (e.g. swapped a couple of words).

Now the app will show a message if no account was found for the
entered recovery key and ask the user to try again.