Workaround iOS bug that crashes app when doing HTTP requests in background
Pine sometimes crashes intermittently when starting it from the background.
This seems to be because it has been doing HTTP requests in the background
and crashed. After a lot of testing and debugging, the only error I've
found is BoringSSL Lower protocol stack error: 53. Altough it didn't
crash the app, this seems related. This is a known bug in iOS that
will be fixed in the next major version, but the workaround is to
delay all HTTP requests that is being performed in background tasks.

If this works, the actual delay could probably be lowered from 1s to
maybe 3ms or so. Let's hope it works 🤞