Don't expire UTXOs
Expiring UTXOs creates problems because it doesn't actually mean that
the transaction will be invalidated. If the user spends one of the
same UTXOs then it will be "canceled" but won't show up as canceled
in the UI. If the user doesn't spend the same UTXOs, then the other
user might still redeem their payment.

So remove the expiration and instead let the user manually cancel
each payment that hasn't been broadcasted.