Redirect to SplashScreen after reset to reload the app
Payments could fail and other unexpected errors could occur if a
new account was created after resetting the app. This was because
the state was invalid and by redirecting to the splash screen, we
can be sure that the app has been reloaded correctly. Why not re-
load the state after the reset? A single point of entry is easier.