Fix bug that crashed app on iPhones that was reset from backup
Resetting or restoring an iPhone from a backup (e.g. when buying a
new phone) restores all data except the keychain, which is where
the mnemonic is stored. This means that the app thinks it's an
existing wallet with a reference to a key in the keychain. This
causes the app to throw an error and then crash.

The fix is to catch the error and reset the app and let the user
recover from an iCloud or manual Pine backup. This could later be
improved to detect if the account was backed up in iCloud and then
automatically recover that mnemonic while keeping all the data.