Only check clipboard when app becomes active instead of every second
In iOS 14 there's a notification showing every time the app accesses
the clipboard, which it does every second. This commit fixes that by
instead only checking the clipboard when the app becomes active.

Drawback: It won't notice addresses copied from within the app until
the app has lost focus and became active again. This can of course be
fixed later, if desired.

Fixes #7